Refactor SQL generating codes into methods of Metrics. This makes Operations work with custom metrics in SQL computation.
PiperOrigin-RevId: 362416635
